pyproject.toml
[project]
name = "graphrag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.116.1",
    "langchain>=0.3.27",
    "langchain-core>=0.3.74",
    "langchain-experimental>=0.3.4",
    "langchain-google-genai>=2.1.9",
    "langchain-neo4j>=0.0.10",
    "python-dotenv>=1.1.1",
    "streamlit>=1.48.1",
    "PyPDF2>=3.0.1",
    "pypdf>=4.0.0",
    "python-multipart>=0.0.6",
    "pandas>=2.0.0",
    "plotly>=6.3.0",
    "pyvis>=0.3.2",
]
